+++ send_voise_message.py
@@ -6,12 +6,12 @@ from airflow.sensors.time_delta import TimeDeltaSensor
 from airflow.exceptions import AirflowException
 from airflow.models import Variable
 from datetime import datetime, timedelta
+from twilio.rest import Client
 import requests
 import os
 import logging
 from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.util.retry import Retry
-from twilio.rest import Client
 
 # Configure logging
 logging.basicConfig(level=logging.DEBUG)
@@ -25,6 +25,7 @@ default_args = {
     "retry_delay": timedelta(minutes=5),
 }
 
+# Twilio Credentials
 TWILIO_ACCOUNT_SID = Variable.get("TWILIO_ACCOUNT_SID")
 TWILIO_AUTH_TOKEN = Variable.get("TWILIO_AUTH_TOKEN")
 TWILIO_PHONE_NUMBER = Variable.get("TWILIO_PHONE_NUMBER")
@@ -32,38 +33,70 @@ TWILIO_PHONE_NUMBER = Variable.get("TWILIO_PHONE_NUMBER")
 if not all([TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN, TWILIO_PHONE_NUMBER]):
     raise ValueError("Twilio credentials are missing. Set them in Airflow Variables.")
 
+# Base directory for storing recordings
 BASE_STORAGE_DIR = "/appz/data"
 
-class TwilioVoiceTranscriber:
-    """A class to handle Twilio voice calls with recording and transcription of the same audio."""
+def make_api_request(url, method="GET", auth=None, retries=3):
+    """Helper function for API requests with timeout and retry logic"""
+    session = requests.Session()
+    retry_strategy = Retry(
+        total=retries,
+        backoff_factor=1,
+        status_forcelist=[500, 502, 503, 504],
+    )
+    adapter = HTTPAdapter(max_retries=retry_strategy)
+    session.mount("http://", adapter)
+    session.mount("https://", adapter)
+    
+    try:
+        if method == "GET":
+            response = session.get(url, auth=auth or (TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN), timeout=10)
+        elif method == "POST":
+            response = session.post(url, auth=auth or (TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN), timeout=10)
+        else:
+            raise ValueError(f"Unsupported method: {method}")
 
-    def __init__(self, account_sid, auth_token, twilio_phone_number, base_storage_dir="/appz/data"):
-        self.client = Client(account_sid, auth_token)
-        self.twilio_phone_number = twilio_phone_number
-        self.base_storage_dir = base_storage_dir
-        self.auth = (account_sid, auth_token)
+        response.raise_for_status()
+        return response
+    except requests.exceptions.RequestException as e:
+        logger.error(f"API request failed: {str(e)}")
+        raise
 
-    def make_api_request(self, url, method="GET", auth=None, retries=3):
-        session = requests.Session()
-        retry_strategy = Retry(
-            total=retries,
-            backoff_factor=1,
-            status_forcelist=[500, 502, 503, 504],
-        )
-        adapter = HTTPAdapter(max_retries=retry_strategy)
-        session.mount("http://", adapter)
-        session.mount("https://", adapter)
-        
-        try:
-            response = session.get(url, auth=auth or self.auth, timeout=10)
-            response.raise_for_status()
-            return response
-        except requests.exceptions.RequestException as e:
-            logger.error(f"API request failed: {str(e)}")
-            raise
+with DAG(
+    "send-voice-message",
+    default_args=default_args,
+    schedule_interval=None,
+    catchup=False,
+    render_template_as_native_obj=True,
+    params={
+        "phone_number": Param("+1234567890", type="string", title="Phone Number"),
+        "message": Param("Hello, please leave a message after the beep.", type="string", title="Message Before Beep"),
+        "need_ack": Param(False, type="boolean", title="Require Acknowledgment"),
+        "call_id": Param(None, type=["string", "null"], title="Call ID"),
+    }
+) as dag:
 
-    def initiate_call(self, phone_number, message, need_ack=True, call_id=None, ti=None):
+    def initiate_call(**kwargs):
+        """
+        Initiate the voice call using Twilio API with recording and transcription.
+        """
+        ti = kwargs["ti"]
+        conf = kwargs["params"]
+        logger.info(f"Received conf: {conf}")
+
+        if not conf.get("phone_number") or not conf.get("message"):
+            logger.error("Missing required parameters: phone_number or message")
+            ti.xcom_push(key="call_outcome", value="Failed")
+            raise ValueError("Missing required parameters: phone_number or message")
+
+        phone_number = conf["phone_number"]
+        message = conf["message"]
+        need_ack = conf.get("need_ack", False)
+        call_id = conf.get("call_id")
+
         logger.info(f"Initiating call to: {phone_number}, need_ack: {need_ack}, call_id: {call_id}")
+
+        client = Client(TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN)
         try:
             if need_ack:
                 twiml = f"""
@@ -81,9 +114,9 @@ class TwilioVoiceTranscriber:
                     <Say>Thank you! Goodbye.</Say>
                 </Response>
                 """
-            call = self.client.calls.create(
+            call = client.calls.create(
                 to=phone_number,
-                from_=self.twilio_phone_number,
+                from_=TWILIO_PHONE_NUMBER,
                 twiml=twiml,
             )
             logger.info(f"Call initiated with SID: {call.sid}")
@@ -96,159 +129,161 @@ class TwilioVoiceTranscriber:
             ti.xcom_push(key="call_outcome", value="Failed")
             raise
 
-    def check_call_status(self, call_sid, call_id=None, need_ack=True, max_attempts=50, delay_seconds=5):
-        import time
+    def check_call_status(**kwargs):
+        """
+        Poll Twilio for call status; raise if still in-progress.
+        """
+        ti = kwargs["ti"]
+        conf = kwargs["params"]
+
+        call_sid = ti.xcom_pull(task_ids="initiate_call", key="call_sid")
+        call_id = conf.get("call_id")
+
         if not call_sid:
-            raise ValueError("Missing call_sid.")
-        for attempt in range(max_attempts):
-            call = self.client.calls(call_sid).fetch()
-            current_status = call.status
-            logger.info(f"Call SID={call_sid}, call ID={call_id}, status={current_status}")
-            if current_status in ["completed", "no-answer", "busy", "failed"]:
-                return current_status
-            else:
-                logger.info(f"Call still in progress, attempt {attempt + 1}/{max_attempts}")
-                time.sleep(delay_seconds)
-        raise AirflowException(f"Call did not complete after {max_attempts} attempts.")
+            raise ValueError("No call_sid found. Did 'initiate_call' fail?")
 
-    def fetch_and_save_recording(self, call_sid, call_id=None, need_ack=True):
+        url = f"https://api.twilio.com/2010-04-01/Accounts/{TWILIO_ACCOUNT_SID}/Calls/{call_sid}.json"
+        response = make_api_request(url)
+        current_status = response.json()["status"]
+        logger.info(f"Call SID={call_sid}, call_id={call_id}, status={current_status}")
+
+        if current_status in ["completed", "no-answer", "busy", "failed"]:
+            ti.xcom_push(key="call_status", value=current_status)
+            if call_id:
+                logger.info(f"Storing Twilio final status in Variable for cross-DAG usage: twilio_call_status_{call_id}")
+                Variable.set(f"twilio_call_status_{call_id}", current_status)
+            return current_status
+        else:
+            raise AirflowException(f"Call is still {current_status}. Retrying...")
+
+    def branch_recording_logic(**kwargs):
+        """
+        If final status is 'completed', proceed to 'fetch_and_save_recording'.
+        Otherwise skip.
+        """
+        ti = kwargs["ti"]
+        final_status = ti.xcom_pull(task_ids="check_call_status", key="call_status")
+
+        logger.info(f"branch_recording_logic sees final_status={final_status}")
+        if final_status == "completed":
+            return "fetch_and_save_recording"
+        else:
+            return "skip_recording"
+
+    def fetch_and_save_recording(**kwargs):
+        """
+        Fetch and save the recording if need_ack=True and call_status='completed'.
+        """
+        ti = kwargs["ti"]
+        conf = kwargs["params"]
+
+        call_sid = ti.xcom_pull(task_ids="initiate_call", key="call_sid")
+        final_status = ti.xcom_pull(task_ids="check_call_status", key="call_status")
+        need_ack = ti.xcom_pull(task_ids="initiate_call", key="need_ack")
+        call_id = conf.get("call_id")
+
+        logger.info(f"fetch_and_save_recording with call_sid={call_sid}, status={final_status}, call_id={call_id}")
+
         if not need_ack:
-            logger.info("need_ack is False, no recording or transcription to process.")
-            return {"message": "No acknowledgment required."}
+            logger.info("need_ack=False, skipping recording download.")
+            ti.xcom_push(key="recording_status", value="No Recording Needed")
+            return {"message": "Recording not needed."}
 
-        recordings = self.client.recordings.list(call_sid=call_sid)
-        if not recordings:
-            logger.info("Recording not found yet.")
-            return {"message": "Recording not available yet.", "transcription": "No transcription available"}
-
-        recording_sid = recordings[0].sid
-        recording_url = f"https://api.twilio.com{recordings[0].uri.replace('.json', '.mp3')}"
-        execution_date = datetime.now().strftime("%Y-%m-%d")
-        dag_name = "send-voice-message-1"
-        save_directory = os.path.join(self.base_storage_dir, dag_name, execution_date)
-        os.makedirs(save_directory, exist_ok=True)
-        file_path = os.path.join(save_directory, f"{call_sid}.mp3")
+        if final_status != "completed":
+            logger.info(f"Cannot fetch recording because final_status={final_status}")
+            ti.xcom_push(key="recording_status", value="Recording Unavailable")
+            return {"message": f"Status={final_status}, no recording."}
 
         try:
-            response = self.make_api_request(recording_url)
-            with open(file_path, "wb") as f:
-                f.write(response.content)
-            logger.info(f"Recording saved at {file_path}")
+            url = f"https://api.twilio.com/2010-04-01/Accounts/{TWILIO_ACCOUNT_SID}/Recordings.json?CallSid={call_sid}"
+            response = make_api_request(url)
+            recordings = response.json().get("recordings", [])
+
+            if not recordings:
+                logger.info("Recording not found yet.")
+                ti.xcom_push(key="recording_status", value="Recording Not Found")
+                raise AirflowException("Recording not found yet.")
+
+            recording_sid = recordings[0]["sid"]
+            recording_url = f"https://api.twilio.com/2010-04-01/Accounts/{TWILIO_ACCOUNT_SID}/Recordings/{recording_sid}.mp3"
+            execution_date = datetime.now().strftime("%Y-%m-%d")
+            dag_name = "send-voice-message"
+            save_directory = os.path.join(BASE_STORAGE_DIR, dag_name, execution_date)
+            os.makedirs(save_directory, exist_ok=True)
+            file_path = os.path.join(save_directory, f"{call_sid}.mp3")
+
+            response = make_api_request(recording_url)
+            if response.status_code == 200:
+                with open(file_path, "wb") as f:
+                    f.write(response.content)
+                logger.info(f"Recording saved at {file_path}")
+                ti.xcom_push(key="recording_status", value="Recording Saved")
+                ti.xcom_push(key="recording_sid", value=recording_sid)
+                if call_id:
+                    Variable.set(f"twilio_recording_file_{call_id}", file_path)
+                    logger.info(f"Set Variable twilio_recording_file_{call_id} to: {file_path}")
+                    ti.xcom_push(key=f"recording_file_{call_id}", value=file_path)
+                return {"message": "Recording downloaded successfully", "file_path": file_path}
+            else:
+                logger.error(f"Failed to download recording. status code={response.status_code}")
+                ti.xcom_push(key="recording_status", value="Recording Failed")
+                raise AirflowException(f"Failed to download recording: {response.status_code}")
+
         except Exception as e:
-            logger.error(f"Failed to download recording: {str(e)}")
-            return {"message": "Failed to download recording", "transcription": "No transcription available"}
+            logger.error(f"Recording fetch error: {str(e)}")
+            ti.xcom_push(key="recording_status", value="Failed")
+            raise
 
-        import time
-        transcriptions = self.client.recordings(recording_sid).transcriptions.list()
-        if transcriptions:
-            transcription = transcriptions[0]
-            for _ in range(10):  # Wait up to 50 seconds
-                transcription = self.client.transcriptions(transcription.sid).fetch()
-                if transcription.status == "completed":
-                    transcription_text = transcription.transcription_text
-                    logger.info(f"Transcribed text for call SID={call_sid}: {transcription_text}")
-                    return {
-                        "message": "Recording downloaded and transcribed successfully",
-                        "file_path": file_path,
-                        "transcription": transcription_text
-                    }
-                elif transcription.status == "failed":
-                    logger.error(f"Transcription failed for call SID={call_sid}")
-                    return {
-                        "message": "Recording downloaded but transcription failed",
-                        "file_path": file_path,
-                        "transcription": "Transcription failed"
-                    }
-                time.sleep(5)
-            logger.info(f"Transcription not completed in time for call SID={call_sid}")
-            return {
-                "message": "Recording downloaded but transcription not completed",
-                "file_path": file_path,
-                "transcription": "Transcription pending"
-            }
-        else:
+    def fetch_transcription(**kwargs):
+        """
+        Fetch transcription and store it if completed; retry if pending up to a limit.
+        """
+        ti = kwargs["ti"]
+        conf = kwargs["params"]
+
+        call_sid = ti.xcom_pull(task_ids="initiate_call", key="call_sid")
+        recording_sid = ti.xcom_pull(task_ids="fetch_and_save_recording", key="recording_sid")
+        call_id = conf.get("call_id")
+
+        url = f"https://api.twilio.com/2010-04-01/Accounts/{TWILIO_ACCOUNT_SID}/Recordings/{recording_sid}/Transcriptions.json"
+        response = make_api_request(url)
+        transcriptions = response.json().get("transcriptions", [])
+
+        if not transcriptions:
             logger.info(f"No transcription available yet for call SID={call_sid}")
+            raise AirflowException("Transcription not yet available; retrying...")
+
+        transcription_url = f"https://api.twilio.com/2010-04-01/Accounts/{TWILIO_ACCOUNT_SID}/Transcriptions/{transcriptions[0]['sid']}.json"
+        transcription_response = make_api_request(transcription_url)
+        transcription = transcription_response.json()
+
+        if transcription["status"] == "completed":
+            transcription_text = transcription["transcription_text"]
+            logger.info(f"Transcribed text for call SID={call_sid}: {transcription_text}")
+            if call_id:
+                Variable.set(f"twilio_transcription_{call_id}", transcription_text)
+                logger.info(f"Set Variable twilio_transcription_{call_id} to: {transcription_text}")
+            ti.xcom_push(key="transcription_status", value="completed")
             return {
-                "message": "Recording downloaded but no transcription available",
-                "file_path": file_path,
-                "transcription": "No transcription available"
+                "message": "Transcription fetched successfully",
+                "transcription": transcription_text
             }
-
-with DAG(
-    "send-voice-message-1",
-    default_args=default_args,
-    schedule_interval=None,
-    catchup=False,
-    render_template_as_native_obj=True,
-    params={
-        "phone_number": Param("+1234567890", type="string", title="Phone Number"),
-        "message": Param("Hello, please leave a message after the beep.", type="string"),
-        "need_ack": Param(False, type="boolean", title="Require Acknowledgment"),
-        "call_id": Param(None, type=["string", "null"], title="Call ID"),  # Added call_id to params
-    }
-) as dag:
-
-    transcriber = TwilioVoiceTranscriber(
-        account_sid=TWILIO_ACCOUNT_SID,
-        auth_token=TWILIO_AUTH_TOKEN,
-        twilio_phone_number=TWILIO_PHONE_NUMBER,
-        base_storage_dir=BASE_STORAGE_DIR
-    )
-
-    def initiate_call_task_func(**kwargs):
-        ti = kwargs["ti"]
-        params = kwargs["params"]
-        transcriber.initiate_call(
-            phone_number=params["phone_number"],
-            message=params["message"],
-            need_ack=params["need_ack"],
-            call_id=params.get("call_id"),
-            ti=ti
-        )
-
-    def check_call_status_task_func(**kwargs):
-        ti = kwargs["ti"]
-        conf = kwargs["params"]
-        call_sid = ti.xcom_pull(task_ids="initiate_call", key="call_sid")
-        call_id = conf.get("call_id")
-        need_ack = ti.xcom_pull(task_ids="initiate_call", key="need_ack")
-
-        if not call_sid:
-            raise ValueError("Missing call_sid. Did 'initiate_call' run?")
-        
-        status = transcriber.check_call_status(call_sid, call_id, need_ack)
-        if call_id:
-            Variable.set(f"twilio_call_status_{call_id}", status)
-            ti.xcom_push(key=f"call_status_{call_id}", value=status)
-        ti.xcom_push(key="call_status_final", value=status)
-        return status
-
-    def branch_based_on_status(**kwargs):
-        ti = kwargs["ti"]
-        final_status = ti.xcom_pull(task_ids="check_call_status", key="call_status_final")
-        return "fetch_recording_task" if final_status == "completed" else "skip_recording"
-
-    def fetch_recording_task_func(**kwargs):
-        ti = kwargs["ti"]
-        conf = kwargs["params"]
-        call_sid = ti.xcom_pull(task_ids="initiate_call", key="call_sid")
-        call_id = conf.get("call_id")
-        need_ack = ti.xcom_pull(task_ids="initiate_call", key="need_ack")
-
-        result = transcriber.fetch_and_save_recording(call_sid, call_id, need_ack)
-        if call_id and "transcription" in result and result["transcription"] not in ["No transcription available", "Transcription failed", "Transcription pending"]:
-            Variable.set(f"twilio_transcription_{call_id}", result["transcription"])
-            logger.info(f"Set Variable twilio_transcription_{call_id} to: {result['transcription']}")
-        ti.xcom_push(key="recording_status", value="Recording Saved" if result["message"].startswith("Recording downloaded") else "Recording Failed")
-        return result
+        elif transcription["status"] == "failed":
+            logger.error(f"Transcription failed for call SID={call_sid}")
+            ti.xcom_push(key="transcription_status", value="failed")
+            return {"message": "Transcription failed", "transcription": "Transcription failed"}
+        else:
+            logger.info(f"Transcription not completed yet for call SID={call_sid}")
+            raise AirflowException("Transcription still pending; retrying...")
 
     # Define Operators
     initiate_call_task = PythonOperator(
         task_id="initiate_call",
-        python_callable=initiate_call_task_func,
+        python_callable=initiate_call,
+        provide_context=True
     )
 
-    wait_task = TimeDeltaSensor(
+    wait_call_status = TimeDeltaSensor(
         task_id="wait_for_call_status",
         delta=timedelta(seconds=30),
         poke_interval=10,
@@ -257,24 +292,36 @@ with DAG(
 
     check_status_task = PythonOperator(
         task_id="check_call_status",
-        python_callable=check_call_status_task_func,
+        python_callable=check_call_status,
+        provide_context=True,
         retries=50,
         retry_delay=timedelta(seconds=5)
     )
 
-    branch_task = BranchPythonOperator(
-        task_id="branch_based_on_status",
-        python_callable=branch_based_on_status,
+    branch_recording_task = BranchPythonOperator(
+        task_id="branch_recording_logic",
+        python_callable=branch_recording_logic,
+        provide_context=True
+    )
+
+    fetch_recording_task = PythonOperator(
+        task_id="fetch_and_save_recording",
+        python_callable=fetch_and_save_recording,
+        provide_context=True
+    )
+
+    fetch_transcription_task = PythonOperator(
+        task_id="fetch_transcription",
+        python_callable=fetch_transcription,
+        provide_context=True,
+        retries=12,  # 60 seconds total (12 retries * 5s retry_delay)
+        retry_delay=timedelta(seconds=5)
     )
 
     skip_recording = DummyOperator(task_id="skip_recording")
 
-    fetch_recording_task = PythonOperator(
-        task_id="fetch_recording_task",
-        python_callable=fetch_recording_task_func,
-    )
-
-    # Set Dependencies
-    initiate_call_task >> wait_task >> check_status_task
-    check_status_task >> branch_task
-    branch_task >> [fetch_recording_task, skip_recording]
+    # Task Dependencies
+    initiate_call_task >> wait_call_status >> check_status_task
+    check_status_task >> branch_recording_task
+    branch_recording_task >> [fetch_recording_task, skip_recording]
+    fetch_recording_task >> fetch_transcription_task
